+++ tests/test_api.py
@@ -622,12 +622,14 @@ class TestElement:
         assert isinstance(bot.obj, User)
         assert bot.obj.type == "bot"
         assert bot.obj.name == "Pytion tests"
+        assert bot.obj.workspace_name == "Yegor's Workspace"
 
     def test_get_myself__from_obj(self, root_page):
         bot = root_page.get_myself()
         assert isinstance(bot.obj, User)
         assert bot.obj.type == "bot"
         assert bot.obj.name == "Pytion tests"
+        assert bot.obj.workspace_name == "Yegor's Workspace"
 
     def test_from_linkto__base(self, no):
         link = LinkTo.create(page_id="878d628488d94894ab14f9b872cd6870")